pyproject.toml
[tool.poetry]
name = "tracker"
version = "0.1.0"
description = ""
authors = ["NerdPraise <praiseajayi2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
asgiref = "3.7.2"
autopep8 = "2.0.4"
django = "5.0.2"
django-cors-headers = "4.3.1"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.1"
flake8 = "7.0.0"
mccabe = "0.7.0"
pycodestyle = "2.11.1"
pyflakes = "3.2.0"
pyjwt = "2.8.0"
python-decouple = "3.8"
pytz = "2024.1"
sqlparse = "0.4.4"
toml = "0.10.2"
ruff = "0.3.0"
psycopg2 = "^2.9.9"
isort = "^5.13.2"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
cache-dir = "~/.cache/ruff"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]
